Version2 The change to the behat.yml file is me just trying to tidy up, you can ignore that.
The changing of "outside the house" to "in the front garden" is just an
example of me consolidating the language for consistency. I don't think
this is always appropriate, the language is always primary. However, in
some cases I think it can actually help the language to be more
consistent on some things.

Finally, I've added a context file which implements
`SnippetAcceptingContext` which means it can be used to generate
some step definitions for me.

Once this is done, I run `bin/behat —append-snippets` to fill the
context file with some template step definitions.