Version1 Ignore the change to AppKernel which was again me thinking of the framework when I shouldn't have been.
I've added my first feature file. It contains some examples of how the
feature 'navigating the house' would work. Looking back, I'm now
thinking this should be broken up. So, enter the house is a feature and
the scenarios should contain examples of that. Overall though, you get
the idea.

This file is the kind of thing that would derive from the conversation
with the client. No code, just language.