Version17 So at this point I'm pretty happy with the domain. All the behat steps pass, the PHPSpec specs too. I have the following:
* A working domain model
* Created from the ubiquitous language
* High level behat stories and lower level PHPSpec tests proving it
* We've not touched any framework code
* We've kept code to a minimum
* We're completely decoupled

If you think about the clean architecture, we have completed that middle
layer, or the logic board on a Raspberry Pi if you will. The next step
is to produce the ports into the domain. The USB and other ports on a
Raspberry Pi if I may continue the analogy.

Therefore, in this commit I've simply added two commands for entering
rooms and exiting rooms. These are two lightweight DTOs that will be
used as a means of transferring messages from our framework into our
domain.