Version19 So as you may have noticed I've been avoiding persistence throughout this. It's a small, simple, contrived example which I didn't want to muddy by adding any means of persistence. I realise this perhaps removes validity to the example, but I don't really care. This is more about how to model a domain with a modelling by example approach, as opposed to the intricacies of database persistence with DDD.
However, I did want a tidier way of dealing with the state of the House.
Building the house in the constructor every time wasn't ideal. So I came
up with this idea of a factory. Because we all know houses come from
factories right? In essence it's more of a singleton, which works for me
here. I'm stil building the house on every request, but at least this
does it in a clear and re-usable manner. It also doesn't get me bogged
down in persistence when it's not my primary aim here.

The final thing to note is that I added some basic handlers to handle
the commands I created earlier. These are likely to change but are fine
for the time being.