Version22 The main thing I did in this commit was to add an event dispatcher. I throw a domain event in my model when entering a room and needed the dispatcher in place to support that.
As I am using Symfony it made sense to me to use the Symfony event
dispatcher. However, it is a bit of a pain because it expects to have a
name and to be `Symfony\Component\EventDispatcher\Event`, as I was
trying to follow the clean architecture and have no outer dependencies I
didn't want to have my event extend
`Symfony\Component\EventDispatcher\Event`. Therefore, I needed to create
an adapter. Going back to the Raspberry Pi analogy made in a previous
version. I needed a serial to usb adapter, or something like that.

I also wanted to use my own Dispatcher interface, because what if I
wanted to stop using the Symfony dispatcher and replace it with
something else? I want to be able to continue using the same interface
but just add a new adapter and change my DI. You can see how I achieved
this from the code within this commit.

Finally, I added a `HandlerInterface` so that I can swap out handlers
using DI should I need to.